+++ source/api_cc/include/commonPT.h
@@ -19,6 +19,6 @@ struct NeighborListDataPT {
   std::vector<int*> firstneigh;
 
  public:
-  void copy_from_nlist(const InputNlist& inlist, unsigned long int& max_num_neighbors);
+  void copy_from_nlist(const InputNlist& inlist, int& max_num_neighbors);
 };
 }  // namespace deepmd
